Use rankpool
rankpool.c contains code to start the rankpool. It
runs for 30 seconds, and its name is set to NULL so
it uses the default rankpool (which is consistent with
not setting the MPI_Info for MPI_Comm_spawn).

The job and submit scripts were also updated to add
nodes for the child job and to launch it.